% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\name{compose_L3}
\alias{compose_L3}
\title{Construct an new-convention L3 filename from consituent segments}
\usage{
compose_L3(mission = c("MODIS", "S3A", "SNPP", "ADEOS", "SEASTAR")[1],
  instrument = c("AQUA", "TERRA", "OLCI", "SEAWIFS", "VIIRS", "OCTS")[1],
  type = c(NA, "ERR", "GAC", "EFR"), date = Sys.Date(),
  level = "L3m", period = "DAY", suite = "CHL",
  product = "chlor_a", nrt = NA, res = "4km", ext = "nc")
}
\arguments{
\item{mission}{character such as 'MODIS'}

\item{instrument}{character such as 'AQUA'}

\item{type}{character such as 'GAC' or NA to skip}

\item{date}{Date one or two elements depending upon the \code{period}}

\item{level}{character such as 'L3m'}

\item{period}{character such as 'DAY'}

\item{suite}{character such as 'CHL' or NA to skip}

\item{product}{character such as 'chlor_a' or NA to skip}

\item{nrt}{character, 'NRT' (near real time) or NA to skip}

\item{res}{character such as '4km' or NA to skip}

\item{ext}{character such as 'nc' or NA to skip}
}
\value{
a character L3 filename
}
\description{
Construct an new-convention L3 filename from consituent segments
}
\seealso{
\href{https://oceancolor.gsfc.nasa.gov/docs/filenaming-convention/}{Ocean Color Filename Convention}
}
