% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Datasets.R
\name{DatasetsRefClass_get_collection}
\alias{DatasetsRefClass_get_collection}
\title{Retrieve the datasets from a dataset collection}
\value{
a list of DatasetRefClass or NULL
}
\description{
Retrieve the datasets from a dataset collection
}

