% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\name{decompose_L3}
\alias{decompose_L3}
\title{Decompose a new-convention L3 filename}
\usage{
decompose_L3(
  x = c("SNPP_VIIRS.20180703.L3m.DAY.OC.chlor_a.4km.nc",
    "SNPP_VIIRS.20180703.L3m.DAY.OC.chl_ocx.4km.nc")
)
}
\arguments{
\item{x}{character vector of one or more filenames}
}
\value{
a tibble of one or more rows (one row for each input filename)
\itemize{
\item{mission character }
\item{instrument character }
\item{type character possibly NA if not present in filename}
\item{date1 Date start date}
\item{date2 numeric or Date end date or NA if none}
\item{level character }
\item{period character }
\item{suite character }
\item{product character }
\item{res character }
\item{nrt character possibly NA if not present in filename}
\item{ext character }
}
}
\description{
Decompose a new-convention L3 filename
}
\seealso{
\href{https://oceancolor.gsfc.nasa.gov/docs/filenaming-convention/}{Ocean Color Filename Convention}
}
