% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct.R
\name{download_direct}
\alias{download_direct}
\title{Download a direct data access file using \code{\link{download.file}}}
\usage{
download_direct(
  filename,
  output_filename,
  output_path = ".",
  base_uri = "https://oceandata.sci.gsfc.nasa.gov/cgi/getfile",
  mode = "wb",
  ...
)
}
\arguments{
\item{filename}{the name of the file to download - required as 'A2016186.L3m_DAY_CHL_chlor_a_4km.nc'}

\item{output_filename}{character the name of the output file, the default is the same as \code{filename}}

\item{output_path}{the file path to save to, the default is to save to the current path "."}

\item{base_uri}{character Oceandata's base uri (default 'http://oceandata.sci.gsfc.nasa.gov/cgi/getfile')}

\item{mode}{character - see \code{\link{download.file}}}

\item{...}{other argments for \code{\link{download.file}}}
}
\value{
0 for success and non-zero otherwise - see \code{\link{download.file}}
}
\description{
Download a direct data access file using \code{\link{download.file}}
}
