% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days.R
\name{get_8days}
\alias{get_8days}
\title{Compute the first or last day of 8-day intervals for a given year.  Handles
leap years when finding the last day of 8-day intervals.}
\usage{
get_8days(year, form = "\%j", when = "first")
}
\arguments{
\item{year}{numeric or character the year(s) to compute}

\item{form}{charcater, by default we return the day of the year in '001' form}

\item{when}{character either 'first' or 'last' to indicate the first or last 
day of the 8-day interval}
}
\value{
a list of character vectors of the last day of the 8-day interval for 
   the years specified with one list element per year.  Be advised that the first
   day of winter for a given year occurs in the prior year.
}
\description{
Compute the first or last day of 8-day intervals for a given year.  Handles
leap years when finding the last day of 8-day intervals.
}
\seealso{
\url{http://oceancolor.gsfc.nasa.gov/BROWSE_HELP/L3/}

Other DAYS: \code{\link{get_monthdays}},
  \code{\link{get_seasondays}}
}

