% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_all_obpg}
\alias{get_all_obpg}
\title{Used to scan all/any days for the listed years}
\usage{
get_all_obpg(
  Product,
  dates = obpg_date(),
  greplargs = NULL,
  verbose = Product$verbose_mode
)
}
\arguments{
\item{Product}{CatalogRefClass}

\item{dates}{tibble as produced by obpg_date}

\item{greplargs}{list of one or more grepl args}

\item{verbose}{logical}
}
\value{
list of zero or more DatasetRefClass objects
}
\description{
Used to scan all/any days for the listed years
}
