% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{get_catalog}
\alias{get_catalog}
\title{Retrieve a catalog}
\usage{
get_catalog(uri)
}
\arguments{
\item{uri}{the URI of the catalog}
}
\value{
ThreddsNodeRefClass or sublass or NULL
}
\description{
Retrieve a catalog
}

