% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days.R
\name{get_monthdays}
\alias{get_monthdays}
\title{Compute the first or last day of the month for a given year.  
Useful for retrieving monthly data stored ont the first day of each month.}
\usage{
get_monthdays(year, form = "\%j", when = c("first", "last")[1])
}
\arguments{
\item{year}{numeric or character the year(s) to compute}

\item{form}{charcater, by default we return the day of the year}

\item{when}{character either 'first' or 'last' to indicate the first or last day of the month}
}
\value{
a list of character vectors of the last day of the months for the years specified
   one list element per year
}
\description{
Compute the first or last day of the month for a given year.  
Useful for retrieving monthly data stored ont the first day of each month.
}
\seealso{
\url{http://oceancolor.gsfc.nasa.gov/BROWSE_HELP/L3/}

Other DAYS: 
\code{\link{get_8days}()},
\code{\link{get_seasondays}()}
}
\concept{DAYS}
