% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days.R
\name{get_seasondays}
\alias{get_seasondays}
\title{Compute the first or last day of seasons for a given year.  
Useful for retrieving seasonal data.}
\usage{
get_seasondays(year, form = "\%j", when = c("first", "last")[1])
}
\arguments{
\item{year}{numeric or character the year(s) to compute}

\item{form}{charcater, by default we return the day of the year}

\item{when}{character either 'first' or 'last' to indicate the first or last day of the season}
}
\value{
a list of character vectors of the last day of the seasons for the years 
   specified with one list element per year
}
\description{
Compute the first or last day of seasons for a given year.  
Useful for retrieving seasonal data.
}
\seealso{
\url{http://oceancolor.gsfc.nasa.gov/BROWSE_HELP/L3/}

Other DAYS: \code{\link{get_8days}},
  \code{\link{get_monthdays}}
}

