% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngm.R
\name{grid_center}
\alias{grid_center}
\title{Compute the grid x,y locations from longitude and latitude}
\usage{
grid_center(theta = 45, lambda = 45, a = 1)
}
\arguments{
\item{theta}{numeric, geographic lat}

\item{lambda}{numeric, geographic lon}

\item{a}{numeric, radius of earth}
}
\value{
two element nmeric vector of x,y
}
\description{
Compute the grid x,y locations from longitude and latitude
}

