% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngm.R
\name{grid_delta}
\alias{grid_delta}
\title{Compute the grid delta}
\usage{
grid_delta(NH = 10, a = 1)
}
\arguments{
\item{NH}{integer, grid spacing}

\item{a}{radius of the earth}
}
\value{
numeric
}
\description{
From The Nested Grid Mode, Norman A. Phillips, April 1979, NOAA Technical Report NWS 22
}
