% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{obpg_date}
\alias{obpg_date}
\title{Convert jjj day numbers to mmdd or the reverse}
\usage{
obpg_date(
  x = Sys.Date() - 1,
  year = format(Sys.Date(), "\%Y"),
  form = c("mmdd", "jjj")[1]
)
}
\arguments{
\item{x}{days of year (1-366), month-days (0101-1231) or dates
\itemize{
\item{day of year, generally a three-digit character, but we try to cast it
      if numeric are passed.  It's best to not pass numerics.}
\item{month-day, generally a four-digit character, but we try to cast it
      if numerics are passed.  It's best to not pass numerics here either.}
\item{Date or POSIXt, if one of these then \code{form} and \code{year} are ignored.}
}}

\item{year}{one or more years to associate with each day of year or month-day.
If shorter than \code{x} then \code{year} is recycled.  Ignored if
\code{x} is a date time class.}

\item{form}{character either 'mmdd', 'jjj'  Ignored if \code{x} is of class
Date or POSIXt}
}
\value{
tibble with these variables
\itemize{
  \item{date Date class objects}
  \item{year character four digit year}
  \item{jjj character three digit day of year}
  \item{mmdd character 4 digit month and day}
}
}
\description{
Convert jjj day numbers to mmdd or the reverse
}
