% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{obpg_query}
\alias{obpg_query}
\title{A basic query function to retrieve one or more datasets}
\usage{
obpg_query(top = "http://oceandata.sci.gsfc.nasa.gov/opendap/catalog.xml",
  platform = "MODISA", product = "L3SMI",
  year = format(as.POSIXct(Sys.Date()), "\%Y"),
  day = format(as.POSIXct(Sys.Date()), "\%j"), what = c("all",
  "most_recent")[1], greplargs = NULL)
}
\arguments{
\item{top}{character, uri of the top catalog}

\item{platform}{character, the name of the platform (MODISA, MODIST, OCTS, SeaWiFS, VIIRS, etc.)}

\item{product}{character, the product type (L3SMI, etc.)}

\item{year}{character or numeric, four digit year(s) - ignored if what = most_recent}

\item{day}{character or numeric, three digit year of day(s) - ignored if what = most_recent}

\item{what}{character, optional filters (defaults to "all" but "most_recent" works)}

\item{greplargs}{list or NULL, if a list the provide two elements,
pattern=character and fixed=logical, which are arguments for \code{grepl} If fixed is FALSE
then be sure to provide a regex for the pattern value.}
}
\value{
list of DatasetRefClass or NULL
}
\description{
A basic query function to retrieve one or more datasets
}
\examples{
\dontrun{
      query <- obpg_query(
         top = 'http://oceandata.sci.gsfc.nasa.gov/opendap/catalog.xml',
         platform = 'MODISA',
         product = 'L3SMI',
         what = 'most_recent',
         greplargs = list(pattern='8D_CHL_chlor_a_4km', fixed = TRUE))
   }
}

