% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct.R
\name{query_direct}
\alias{query_direct}
\title{Query a direct data download webpage such as 
\url{http://oceandata.sci.gsfc.nasa.gov/MODIS-Aqua/Mapped/Daily/4km/chlor_a/2016}}
\usage{
query_direct(base_uri = "https://oceandata.sci.gsfc.nasa.gov",
  mission = "MODIS-Aqua", level = "Mapped", freq = "Daily", res = "4km",
  param = "chlor_a", year = 2016, alt_uri = NULL)
}
\arguments{
\item{base_uri}{character Oceandata's base uri (default 'http://oceandata.sci.gsfc.nasa.gov')}

\item{mission}{character one mission listed at base_uri (default 'MODIS-Aqua')}

\item{level}{character level of processing is mission dependent.  See 
\url{http://oceandata.sci.gsfc.nasa.gov/MODIS-Aqua} for an example. (default 'Mapped')}

\item{freq}{character (or NULL, "" or NA)  Some mission levels are processed at various
freqencies.  See \url{http://oceandata.sci.gsfc.nasa.gov/MODIS-Aqua/Mapped}.  The default ('Daily')
assumes there is a choice, but skip this by setting to \code{freq} to \code{NULL}, 
\code{NA} or \code{""}.}

\item{res}{character (or NULL, "" or NA) resolution.  Defaults to '4km' but you can skip this 
by setting to \code{freq} to \code{NULL}, \code{NA} or \code{""}.}

\item{param}{character the name of the parameter}

\item{year}{numeric (or character) year (2016 default)}

\item{alt_uri}{character or NULL.  Use this to specify your own full uri which ignores
other parameters.}
}
\value{
data.frame or NULL
}
\description{
This is fairly centric to 'Mapped' level data but some flexibility is provided to 
access other levels of data.
}

