% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Service.R
\docType{class}
\name{ServiceRefClass-class}
\alias{ServiceRefClass}
\alias{ServiceRefClass-class}
\title{An Service representation that subclasses from ThreddsNodeRefClass}
\description{
An Service representation that subclasses from ThreddsNodeRefClass
}
\section{Fields}{

\describe{
\item{\code{name}}{character}

\item{\code{serviceType}}{character}

\item{\code{base}}{character base url}
}}
\section{Methods}{

\describe{
\item{\code{initialize(x)}}{x may be url or XML::xmlNode}

\item{\code{show(prefix = "")}}{show the content of the class}
}}
\seealso{
Other Thredds: \code{\link{DatasetRefClass}},
  \code{\link{DatasetRefClass-class}};
  \code{\link{DatasetsRefClass}},
  \code{\link{DatasetsRefClass-class}};
  \code{\link{ThreddsNodeRefClass_GET}};
  \code{\link{ThreddsNodeRefClass_get_url}};
  \code{\link{ThreddsNodeRefClass_unames}};
  \code{\link{ThreddsNodeRefClass}},
  \code{\link{ThreddsNodeRefClass-class}}
}

