% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Thredds.R
\docType{class}
\name{ThreddsNodeRefClass-class}
\alias{ThreddsNodeRefClass}
\alias{ThreddsNodeRefClass-class}
\title{An base representation that  other nodes subclass from}
\description{
An base representation that  other nodes subclass from
}
\section{Fields}{

\describe{
\item{\code{url}}{character - possibly wrong but usually right!}

\item{\code{node}}{XML::xmlNode}
}}
\section{Methods}{

\describe{
\item{\code{initialize(x)}}{x may be url or XML::xmlNode}
}}
\seealso{
Other Thredds: \code{\link{DatasetRefClass}},
  \code{\link{DatasetRefClass-class}};
  \code{\link{DatasetsRefClass}},
  \code{\link{DatasetsRefClass-class}};
  \code{\link{ServiceRefClass}},
  \code{\link{ServiceRefClass-class}};
  \code{\link{ThreddsNodeRefClass_GET}};
  \code{\link{ThreddsNodeRefClass_get_url}};
  \code{\link{ThreddsNodeRefClass_unames}}
}

