% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Thredds.R
\name{ThreddsNodeRefClass_unames}
\alias{ThreddsNodeRefClass_unames}
\title{Retrieve a vector of unique child names}
\value{
a vector of unique children names
}
\description{
Retrieve a vector of unique child names
}
\seealso{
Other Thredds: \code{\link{DatasetRefClass}},
  \code{\link{DatasetRefClass-class}};
  \code{\link{DatasetsRefClass}},
  \code{\link{DatasetsRefClass-class}};
  \code{\link{ServiceRefClass}},
  \code{\link{ServiceRefClass-class}};
  \code{\link{ThreddsNodeRefClass_GET}};
  \code{\link{ThreddsNodeRefClass_get_url}};
  \code{\link{ThreddsNodeRefClass}},
  \code{\link{ThreddsNodeRefClass-class}}
}

