% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpgcrawler-package.R
\docType{package}
\name{obpgcrawler}
\alias{obpgcrawler}
\title{obpgcrawler - programmatic access to Ocean Color OpeNDAP servcies.}
\description{
A crawler for programmatically working with the OpeNDAP data offerings from
}
\examples{
\dontrun{ Top <- threddscrawler::get_catalog(platform = 'MODISA', verbose = TRUE) }
}
